Get up to date

+++ dlib/general_hash/murmur_hash3.h
@@ -320,7 +320,7 @@ namespace dlib
     inline std::pair<uint64,uint64> murmur_hash3_128bit ( 
         const void* key, 
         const int len,
-        const uint32 seed = 0
+        const uint64 seed = 0
     )
     {
         const uint8 * data = (const uint8*)key;
